Refactor 30th minute cron job to class
